Add usb_keyboard/Key
The key press should be sent only if the key is currently not pressed.
The key release should be sent only if no other switch is sending the
keycode.

This commit solves an issue when the same keycode is assigned to
multiple keys. The press/release events would be sent for each switch,
which did not match the behavior observed when multiple keyboards are
connected to a device (i.e. pressing 'a' on keyboard1, then 'a' on
keyboard2, the 'a' char is pressed until both key are released).