Change time management
The Universe now has it's own time, stored in Timer.
This should help having the expected behavior when
an input takes a long time to process.
The Timer starts at the time of the input (instead
of the time the input is processed), and time moves
forward based on the timestamp of the next Event.